rest api

+++ src/main/java/com/manisha/java/retailapp_project/repository/ShopRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ShopRepository extends JpaRepository<Shop, Long> {
+public interface ShopRepository extends JpaRepository<Shop, Integer> {
     List<Shop> findByName(String name);
 
     //public findAll()
